\name{BayesCompRisk-package}
\alias{BayesCompRisk-package}
\alias{BayesCompRisk}

\docType{package}

\title{
 Bayesian inference on parametric and non-parametric competing risks.
}

\description{
This package allows users to carry out Bayesian inference on age-specific competing risks for cause of death analyses. It includes tools for data checking, model diagnostics, and outputs such as plots of goodness of fit, cumulative incidence, and cause-specific hazards.
}

\details{
\tabular{ll}{
	Package: \tab BayesCompRisk \cr
	Type: \tab Package\cr
	Version: \tab 1.0.0\cr
	Date: \tab 2026-05-07\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}

FILL UP DETAILS 
}

\author{Fernando Colchero \email{fernando_colchero@eva.mpg.de}}

\references{
FILL UP
}