\name{BayesCR}
\alias{BayesCR}
\alias{BayesCR.default}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{
 BayesCR(object, \dots)

\method{BayesCR}{default} (object, niter = 11000, burnin = 1001, thinning = 20,  nsim = 4, ncpus = 4, UPDJUMP = TRUE, jumpSD = NULL)

}

 \arguments{
	\item{object }{FILL UP}

	\item{niter }{FILL UP}

	\item{burnin }{FILL UP}

	\item{thinning }{FILL UP}

	\item{nsim }{FILL UP}

	\item{ncpus }{FILL UP}

	\item{UPDJUMP }{FILL UP}

	\item{jumpSD }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{fernando_colchero@eva.mpg.de}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

