\name{BayesCompRisk-package}
\alias{BayesCompRisk-package}
\alias{BayesCompRisk}

\docType{package}

\title{
 FILL UP
}

\description{
 FILL UP
}

\details{
\tabular{ll}{
	Package: \tab BayesCompRisk \cr
	Type: \tab Package\cr
	Version: \tab 1.0.0\cr
	Date: \tab 2025-06-08\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}

FILL UP DETAILS 
}

\author{Fernando Colchero \email{fernando_colchero@eva.mpg.de}}

\references{
FILL UP
}