\name{plot.BayesCR}
\alias{plot.BayesCR}
\alias{plot.BayesCR}
\alias{print.BayesCR}
\alias{summary.BayesCR}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{
 \method{plot}{BayesCR} (x, type = "traces", noCIs = FALSE, logMort = FALSE,  ...)

\method{print}{BayesCR} (x, ...)

\method{summary}{BayesCR} (object, ...)

}

 \arguments{
	\item{x }{FILL UP}

	\item{object }{FILL UP}

	\item{type }{FILL UP}

	\item{noCIs }{FILL UP}

	\item{logMort }{FILL UP}

	\item{\dots }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{fernando_colchero@eva.mpg.de}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

